Decrease badge cache length.
Loading times should not be an issue since I've implemented echo.js and badgen seems to be quite performant.

The active projects cache has been decreased to the minimum of 5 minutes while the archived projects cache has been decreased to an hour as I do not expect to be updating them frequently.